Removed school data service, replaced national score with n/a when not found and renamed method sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId to sqlFindResultsBySchool
