+++ admin/services/data-access/result.data.service.js
@@ -3,6 +3,7 @@
 const resultDataService = {}
 const sqlService = require('./sql.service')
 const TYPES = require('tedious').TYPES
+const R = require('ramda')
 
 /**
  * Find pupils with scores
@@ -10,7 +11,7 @@ const TYPES = require('tedious').TYPES
  * @param {Number} checkWindowId
  * @returns {Promise<*>}
  */
-resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId = async (schoolId, checkWindowId) => {
+resultDataService.sqlFindResultsBySchool = async (schoolId, checkWindowId) => {
   const sql = `
     SELECT
     p.foreName,
@@ -48,6 +49,35 @@ resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId
     }
   ]
   return sqlService.query(sql, params)
+}
+
+/**
+ * Find school score based on school id and check window id
+ * @param {Number} schoolId
+ * @param {Number} checkWindowId
+ * @returns {Promise<*>}
+ */
+resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId = async (schoolId, checkWindowId) => {
+  const sql = `
+  SELECT score
+  FROM mtc_admin.schoolScore
+  WHERE school_id = @schoolId
+  AND checkWindow_id = @checkWindowId
+  `
+  const params = [
+    {
+      name: 'checkWindowId',
+      value: checkWindowId,
+      type: TYPES.Int
+    },
+    {
+      name: 'schoolId',
+      value: schoolId,
+      type: TYPES.Int
+    }
+  ]
+  const result = await sqlService.query(sql, params)
+  return R.head(result)
 }
 
 module.exports = resultDataService
