+++ admin/services/result.service.js
@@ -1,5 +1,4 @@
 const resultDataService = require('../services/data-access/result.data.service')
-const schoolScoreDataService = require('../services/data-access/school-score.data.service')
 
 const resultService = {}
 
@@ -16,7 +15,7 @@ resultService.getPupilsWithResults = async (schoolId, checkWindowId) => {
   if (!checkWindowId) {
     throw new Error('check window id not found')
   }
-  return resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId(schoolId, checkWindowId)
+  return resultDataService.sqlFindResultsBySchool(schoolId, checkWindowId)
 }
 
 /**
@@ -32,7 +31,7 @@ resultService.getSchoolScore = async (schoolId, checkWindowId) => {
   if (!checkWindowId) {
     throw new Error('check window id not found')
   }
-  const schoolScore = await schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId(schoolId, checkWindowId)
+  const schoolScore = await resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId(schoolId, checkWindowId)
   if (!schoolScore || !schoolScore.score) {
     throw new Error(`no school score record is found or no score is set for school id: ${schoolId} and check window id: ${checkWindowId}`)
   }
