+++ admin/spec/back-end/service/result.service.spec.js
@@ -2,13 +2,12 @@
 /* global describe, it, expect spyOn fail */
 
 const resultDataService = require('../../../services/data-access/result.data.service')
-const schoolScoreDataService = require('../../../services/data-access/school-score.data.service')
 const resultService = require('../../../services/result.service')
 
 describe('result.service', () => {
   describe('getPupilsWithResults', () => {
-    it('calls sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId when school id and check window id are provided', async () => {
-      spyOn(resultDataService, 'sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId')
+    it('calls sqlFindResultsBySchool when school id and check window id are provided', async () => {
+      spyOn(resultDataService, 'sqlFindResultsBySchool')
       const checkWindowId = 1
       const schoolId = 2
       try {
@@ -16,10 +15,10 @@ describe('result.service', () => {
       } catch (error) {
         fail()
       }
-      expect(resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId).toHaveBeenCalled()
+      expect(resultDataService.sqlFindResultsBySchool).toHaveBeenCalled()
     })
     it('throws an error if check window id is not provided', async () => {
-      spyOn(resultDataService, 'sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId')
+      spyOn(resultDataService, 'sqlFindResultsBySchool')
       const checkWindowId = undefined
       const schoolId = 2
       try {
@@ -28,10 +27,10 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe('check window id not found')
       }
-      expect(resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
+      expect(resultDataService.sqlFindResultsBySchool).not.toHaveBeenCalled()
     })
     it('throws an error if school id is not provided', async () => {
-      spyOn(resultDataService, 'sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId')
+      spyOn(resultDataService, 'sqlFindResultsBySchool')
       const checkWindowId = 1
       const schoolId = undefined
       try {
@@ -40,12 +39,12 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe('school id not found')
       }
-      expect(resultDataService.sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
+      expect(resultDataService.sqlFindResultsBySchool).not.toHaveBeenCalled()
     })
   })
   describe('getSchoolScore', () => {
     it('calls sqlFindSchoolScoreBySchoolIdAndCheckWindowId when school id and check window id are provided and returns a valid score number', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue({ id: 1, score: 6 })
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue({ id: 1, score: 6 })
       const checkWindowId = 1
       const schoolId = 2
       let schoolScore
@@ -54,11 +53,11 @@ describe('result.service', () => {
       } catch (error) {
         fail()
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
       expect(schoolScore).toBe(6)
     })
     it('throws an error if check window id is not provided', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
       const checkWindowId = undefined
       const schoolId = 2
       try {
@@ -67,10 +66,10 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe('check window id not found')
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
     })
     it('throws an error if school id is not provided', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
       const checkWindowId = 1
       const schoolId = undefined
       try {
@@ -79,10 +78,10 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe('school id not found')
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).not.toHaveBeenCalled()
     })
-    it('throws an error if sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId returns an error', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue(Promise.reject(new Error('error')))
+    it('throws an error if sqlFindResultsBySchool returns an error', async () => {
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue(Promise.reject(new Error('error')))
       const checkWindowId = 1
       const schoolId = 2
       try {
@@ -91,10 +90,10 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe('error')
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
     })
-    it('throws an error if sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId returns an empty object', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
+    it('throws an error if sqlFindResultsBySchool returns an empty object', async () => {
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId')
       const checkWindowId = 1
       const schoolId = 2
       try {
@@ -103,10 +102,10 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe(`no school score record is found or no score is set for school id: ${schoolId} and check window id: ${checkWindowId}`)
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
     })
-    it('throws an error if sqlFindPupilsWithScoresAndAttendanceBySchoolIdAndCheckWindowId returns no score property', async () => {
-      spyOn(schoolScoreDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue({ id: 1 })
+    it('throws an error if sqlFindResultsBySchool returns no score property', async () => {
+      spyOn(resultDataService, 'sqlFindSchoolScoreBySchoolIdAndCheckWindowId').and.returnValue({ id: 1 })
       const checkWindowId = 1
       const schoolId = 2
       try {
@@ -115,7 +114,7 @@ describe('result.service', () => {
       } catch (error) {
         expect(error.message).toBe(`no school score record is found or no score is set for school id: ${schoolId} and check window id: ${checkWindowId}`)
       }
-      expect(schoolScoreDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
+      expect(resultDataService.sqlFindSchoolScoreBySchoolIdAndCheckWindowId).toHaveBeenCalled()
     })
   })
 })
