+++ admin/views/results/view-results.ejs
@@ -21,7 +21,7 @@
       </dd>
       <dd>
         <h3>National Average</h3>
-        <div class="font-xlarge"><%= nationalScore ? nationalScore: '0.00' %></div>
+        <div class="font-xlarge"><%= nationalScore ? nationalScore: 'N/A' %></div>
       </dd>
     </dl>
   </div>
